procesador-pdf/requirements.txt
2captcha-python==1.5.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
Appium-Python-Client==4.4.0
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
Faker==33.1.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-dialogflow==2.34.0
googleapis-common-protos==1.66.0
gpt4all==2.8.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
lxml==5.3.0
macholib==1.16.3
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.52.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
playwright==1.48.0
propcache==0.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pyee==12.0.0
pygame==2.6.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
PyTempMail==0.1
python-dateutil==2.9.0.post0
pytz==2024.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.11.2
selenium==4.27.1
service-identity==24.2.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
temp-mail==0.2
tldextract==5.1.2
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
twilio==9.3.6
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
w3lib==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.16.0
zope.interface==7.1.1
